Adding "Math Addons"
Ok hear me out, I know theres already an extension for operators and for more math blocks, but these blocks literally dont exist like this or at all. 
I'll be adding more and more functionality for a variety of blocks. It looks like little amount of blocks but theres a lot of dropdowns